add interrupt for the QWERTYUS keybord and a basic handler for it

+++ src/math.h
@@ -1,7 +1,7 @@
 #pragma once
 
 typedef struct {
-	unsigned int X;
-	unsigned int Y;
+	long int X;
+	long int Y;
 } Point;
